Readability and coverage re-writes
Adds a lintr configuration (run `lintr::lint_package()`) to aid in consistent
coding. Most files were re-written to comply with the lintr specifications
added in the .lintr configuration file.

Replaces the patch to `stats::update` with `listwise_delete()`. This function
will notify when cases with missing values are detected, refit the model, and
emit a message with the equivalent call. It is called automatically by both
`supernova()` and `generate_models()`.

Most (if not all) of the documentation is now written in the Roxygen/Markdown
hybrid format, which is easier to read.

`generate_models()` no longer emits a warning about Type 3 SS being
incompatible with `anova()`. The models are now fit with `lm.fit()` and can
be directly compared.

Adds new `frm_` utility functions and a full test suite for `utils-formula.r`

Adds lowercase versions of estimate extractors: pre, ssr, ssm, sse, b0, b1, f